Rewrote the entire chain.
The pipeline is now heavily reliant on the REQUEST object being passed around instead of different objects being passed and worked on. get_file() and .py endings still need to be looked over. But overall, events are now generated on each step so that external users can follow up on where the data is in the pipeline. Finally, the response will be sent to the requester or be passed to user implemented hooks.