Adding HTTPS support.
Wrote a mockup of PyOpenSSL's `SSL` library. It's taken from the branch `client-cert-auth`. Where PyOpenSSL is favored over `ssl`. However, since we're not depending on external libs (unless opt in), we have to emulate the behavior of `PyOpenSSL.SSL` by using `ssl`. Tested on a \*Nix machine connecting from chrome, and seams to work.