Added two new configuration options.
vhost's *(currently only vhosts)* now supports module importing with partial live reload of code in memory.

The second being a stub for `proxy` support. We'll need to implement a pipe between two endpoints that can talk to each other un-filtered.. after TLS setup. So it's a stub for now, but the option is there.

the `vhost.module = /path/to/script.py` support comes in two flavors. One being a default `on_request` event function, if that is present in the module, that is called by default. If a `$slimHTTP.instances['addr:port'].route('/', vhost='name.com')` is given, that overrides the default and handles the given route/request.