Fixes #8
`proxy` and `module` definitions should *probably* end the request. And not allow for  continuing to the normal `GET` and `POST` functions. However, I can't decide if this behavior is desired that the web-server first tries the `module` definition, and then tries the `GET` function-definition.

But this turns off the later. Making sure that webroots etc never default back to the default `web-root` definition.

TODO: Instead of defaulting back to the default `web-root`, perhaps check if there's a definition for `web-root` as well as `module` for a vhost, and if that's the case then perform the `GET` call? Perhaps overly complicated and not so useful. So saving it in a commit thought : )